Initiate rest lengths
When two masses are connected, compute the length in between and store it as the spring rest length.

This will be utilized when we simulate the system, to check if current distance is different than rest length. (Note that for performance you may opt to use the squared distances in future)