Fix spring force direction
The spring force direction was facing the other way around. Also, note that PyVista's timer callbacks are overriding the assert functions, when assert fails the timer goes back to step 0 (which I could only catch with lucky if statements to print the force is applied in the beginning. Who could've predicted the timer will go back to step zero because an assert has failed?)