Fix velocity update
Velocity updates should include the time step we're using, add the division by dt to the mass velocity updates.

With this fix, the mass-spring simulation works as expected for the first time! However this is just the most trivial case we're observing. We should change the parameters and connectivity to further develop the toy example.