Make mouse mode as initial creation mode
When the app is launched the mouse mode should be pan-zoom, instead of creation modes. So that we user randomly clicks on the screen there won't be any new objects appearing, instead user needs to first select one of the creation objects to start creating.

This modification is done to easily review the mass-spring tests we have set up initially. We should be able to move camera around without adding new mass particles to the system initially.