Add id for Bone class
To keep track of the given matrix data indices and our skeleton classes, add self.idx property to Bone class that will represent the index of the given joint locations.

Note that the joint and bone names are misleading. By design choice, I've represented the joints given by SMPL data as Bones, that have 2 points (one is the endpoint of parent bone, and the other is the joint location given in the matrix). With this design we're able to offset bones that is intended to be used when we add springs to rig, we may need to add zero-length springs that are like zero-length bones in the rig.

Alternatively we could assume all bones are connected, and use large rigid bones to represent the offset, and attach zero-length bone at the end.