Rename function
The function name get_bone_tuple_locations() could be misleading if it is tried to be used after skinning. Calling forward kinematics function (pose function) should return the posed bone vertices, otherwise this function is for the rest bone locations. To eliminate that confusion the function is renamed to get_rest_bone_locations()