Fix rotation multiplication
It turns out our way of multiplying quaternions caused problems because and identity quaternion is [0 0 0 1] but if we directly multiply this numpy array with another numpy array quaternion, we just get identity for all rotations.

So as a solution, use scipy library's rotations to handle quaternion mulitplication.

We can further clean up the code by reducing the number of quat to Rotation conversions we do.