Apply small translation at every X frame
Apply small translation (force) to the fixed point of zero-length spring at every certain step. This is done to simulate the effects of skeletal animation that is updated almost at every frame.

When the translation is small, the point mass jiggles well. However when there are large steps, the jiggling get much larger that might be undesirable.