Rename spring_rig_helper to helper_rig
Previously we were using 3 different names to refer a helper rig. To be more descriptive, change SpringRigContainer to HelperBoneHandler because this module is responsible for handling the relation between a rigid rig and a mass spring system provided which bones are the helper bones.

Rename spring_rig_helper.py to helper_rig.py to support the naming convention between the tests and the module.