Rename skeleton.bones as skeleton.rest_bones
The skeleton class is intended to hold the rest positions of the joints, and when requested it can return the posed positions of these bone hierarchy. However, it never actually overrides the rest positions that is to be utilized while skinning the T-pose meshes.

Since we have this design, the skeleton bones[] should be renamed to rest_bones[] that makes the design explicit, i.e. to access the posed joint locations one should call the pose_bone(), one should not try to access the posed bone locations through bones[ ] (I tried to do that while trying to extend the Skeleton class)