Set default dscale to 1.0
The dscale parameters are not dependable parameters. They help to slow down the velocities and dampen forces further, but they might be misleading when the actual stiffness and mass parameters are set to be unstable. So change dscale defaults to 1.0 if someone does not want to define them at all.