Rename helper_indices to helper_idxs
Practicing naming by shortening the names without creating a fuzzy meaning. Indices are very common variable name I use, I could switch to "idxs" as "idx" is also commonly used for index, but only using helper_idx would be confusing because it looks like helper_idx is a single int whereas it is actually an array of ints. So I preferred to call it idxs.