Add point springs to the tip and limit bone length
Previously the helper bone was the spring itself. Now it can be added to the tip of the bone.

Additionally, we want the helper bone to not scale if the tip is attached to the spring. So limit the mass movement to the helper length. This stabilized the system though the jiggle is more of a circular motion.