Close plotter after assert raised
Try and catch assertion errors during our simulation execution. When an error is raised, close the plotter and raise the error.

Before that we had to restart the kernel because we couldn't close plotter when an assert fails (or asserts would be silenced by PyVista). This allows us to trace what went wrong.