Remove init_pose() dependency
So we needed to call this function before the animation because self.prev_bone_locations would be set to rest pose in the helper rig. Instead, set it to None in the initialization and set it to the first pose if it's None during the update() call.

With this, we don't have to deal with this function during tests as it is internally handled. In addition, we could remove the repeated lines at the update() by calling init_pose() internally.