Fix extrema indices
The previous implementation unnecessarily complicated the algorithm. Change n_zigzags parameter to represent number of zigzags i.e. up and downs, instead of representing the number of extrema (I mean this makes more sense for n_zigzag name anyway, and it's more intuitive).

This approach simplified the function a lot, so we're now ready to implement the crux of this algorithm!