Remove erronous offset
The offset of the bone does not make sense when you're not translating the whole bone. "self.t" parameter was for the offset already. This issue should be addressed later when we have a working chain of helper bones so that we can sophisticate the things. However, for our application it might be safe to limit the scope with the connected bones, without an offset.