Fix helper chain forward kinematics
When added as a chain, helper bones were not following the movement of the skeleton properly. The issue is rooted by the difference vector we're taking, that I unwittingly put inside in a for loop, but the loop is updated for evey helper bone, so that the bones that aren't the first bone were receiving faulty data. Fix the problem by taking that difference vector out (line 246/251)

Also I added some comments to aid the implementation of the next step: make the helper bone head locations follow the simulated.