Test spring bones
Test spring bones instead of point springs attached to helper bones. It's more jiggly as expected. But the fixed scale doesn't seem to be working when set to true (that'S probably because we adjust the start locations *after* rescaling bones, that's not good, we can try to do it simultaneously as we go through bones).