Add more settings to file name
Since we also have boolean variables that indicate whether a bone is a spring or it has a point spring at the tip, or a bone will preserve its length throughout the simulation or not, it's worth to add these settings to the output file name just for the future reference. Though it's not really easy to guess the parameters by just looking, we could print those settings on the screen that'd ease the inspection.