Create spring_lattice_demo.py
Copy the lattice test to the demo files as it's now ready to be demonstrated. This shouldn't be done like this though, I'm thinking about another way to gather demo files instead of copying the test files because if we're just going to copy the test files to the demo, then what's the actual purpose of test files? The test files should test the edge cases, they aren't here only for develop the successful case for demonstration.