Inspect optimal rotations (failure case)
Add optimal rotation and translation obtained from skeleton locations to the skinning framework. To test the helper bones, we first need to bind them with weights so it'll be my next goal.

But first, we observe in this test that optimal rotation is not optimal at all, it introduces extra 180degrees rotation along the bone axis. We need to discard this algorithm and use something else.