Update file hierarchy
As the project grows, I needed more modularized approach to organize the files I'm working with. I've separated the source files from demo and test files. Inside source, there are render, utils and models modules. This version seems to be cleaner now to work with however it's not the ideal.

Main struggle was to handle imports correctly because as we've separated tests into another folder that's outside of our source code, it was tricky to set the imports (relative imports saved the day).