Fix bulging weights
We weren't normalizing the weights in the system that caused translation-like
artifacts in the skinning.
+ Provide a function to normalize weights
+ Fix axis sum bug at _assert_normalized_weights()

Now the skinning works as expected! Next, we should provide transformations for the spring bones to transfer their motion on the surface. It's not a trivial task though. Initially we'll just check the translations but then we shall tackle rotations and scale as well.