Refactor transformations computation
The bone transformations are computed within the rig handler. There
are several ways to compute a transformation of a bone in the rig. In our
setting, we've opted to use optimal rigid motion algorithm to compute them
however we've observed that this approach fails when directly used to
transform a volumetric set of points. So we need another algorithm that doesn't
create the artifacts we get in this algorithm.

To do so, I've refactored the computation of optimal rigid motion into another function. There's still the option to call SVD-based transformations by setting algorithm parameter to "SVD". We can now provide more algorithms to compute the transformation mapping of the bones that can be feed into skinning (also this is a problem to be researched for simulating the bones).