Update linalg_utils.py
Rename the function to be more descriptive. "get_transform_mats" is confusing because the input rotations are expected to be quaternions but then there is another function that can compose a transformation matrix... So I've just renamed this one but these functions can also be merged or clearly separated.

Also fix the assert that fails when a numpy float is given. This version more explicitly checks the types and handles accordingly.