Fix transformation chain
The previous commits were failing in some cases. I've found the problem that we weren't chaining the transformations correctly. Now the transformations are applied one by one order and it produces the expected results, i.e. both endpoints of the source bone can be mapped to the target bone with the returned single transformation matrix.