Collapse interface
I thought we could make a single call in skinning module to compute the positions of the bones. However, this reduces the flexibility in the code and adds unnecessary dependencies without much benefits.

So for example, I want to use SMPL skeleton that is pre-computed already. I should be able to feed the skeleton bones locations to the Helper Rig, whose job is to dynamically update the bone locations given the current locations (next we'll decouple Skeleton and Helper Rig class to do that). The function call at skinning currently adds an extra step to pose the bones without providing any further benefits. So I removed the class to implement the decoupling better.