Move data scripts to src/data
src folder should include all the source scripts, data folder should not include any scripts because it's meant to hold the data, what we provide in the github folders will have an empty data folder and the user should download the related dataset to that folder. Therefore, it's best to move the scripts to the src/data that is dedicated to data-related scripts.