Add warning
Importing .obj to blender might cause vertex loss or change in vertex indices. I was losing 1 vertex and some of the vertex indices were different, so I've added a solution to that in the first step. Unfortunately this caused me to repaint the weights but for the future .obj's I'll follow this procedure.