Print total error metrics
After showing the results, print the total errors. At this stage our error seems to be larger than SMPL despite the fact that we have more jiggling. This should be addressed with better rig and associated weights to see if we can lower it down (ideally we should optimize these parameters)