Separate render and simulation
I've splitted the smpl demo file into two, separating offline simulation part from the render. Right now there are still unnecessary parts to be refactored. I needed to split it to extract simulation variables in a file, and then run the simulations for CPBD, and finally visualize them in a batch.