Standardize helper mode
Now spot rig with helper bones share the same data structure to be used in Wu et al. and LBS too, so that the comparison is fair. Previously I was only adding the helper bones to only our demo but then the time measurements are not valid anymore. Plus, the tetmesh is used for LBS and our skinning too  for consistent and fair comparison again.